--- jormungandr/src/leadership/epoch_parameters.rs
@@ -1,15 +1,0 @@
-use crate::blockcfg::{Epoch, Ledger, LedgerParameters, LedgerStaticParameters};
-
-/// parameters passed to the leadership [`Process`]
-/// and then broadcasted to every leader [`Task`].
-///
-/// [`Process`]: ./struct.Process.html
-/// [`Task`]: ./struct.Task.html
-pub struct EpochParameters {
-    pub epoch: Epoch,
-    pub ledger_static_parameters: LedgerStaticParameters,
-    pub ledger_parameters: LedgerParameters,
-
-    pub time_frame: chain_time::TimeFrame,
-    pub ledger_reference: Ledger,
-}
